init submission 1
