--- OpenMusic API V3 Test.postman_environment.json
@@ -1,234 +1,0 @@
-{
-	"id": "a96dc807-d100-446f-947c-1e38d834ff4a",
-	"name": "OpenMusic API V3 Test",
-	"values": [
-		{
-			"key": "host",
-			"value": "localhost",
-			"enabled": true
-		},
-		{
-			"key": "port",
-			"value": "5000",
-			"enabled": true
-		},
-		{
-			"key": "newTitle",
-			"value": "Kenangan Mantan",
-			"enabled": true
-		},
-		{
-			"key": "newYear",
-			"value": "2021",
-			"enabled": true
-		},
-		{
-			"key": "newPerformer",
-			"value": "Dicoding",
-			"enabled": true
-		},
-		{
-			"key": "newGenre",
-			"value": "Indie",
-			"enabled": true
-		},
-		{
-			"key": "newDuration",
-			"value": "120",
-			"enabled": true
-		},
-		{
-			"key": "badSongPayloads",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "currentBadSongPayload",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "songId",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "updateTitle",
-			"value": "Bintang Di Surga",
-			"enabled": true
-		},
-		{
-			"key": "updateYear",
-			"value": "2003",
-			"enabled": true
-		},
-		{
-			"key": "updatePerformer",
-			"value": "Peterpan",
-			"enabled": true
-		},
-		{
-			"key": "updateGenre",
-			"value": "Pop",
-			"enabled": true
-		},
-		{
-			"key": "updateDuration",
-			"value": "240",
-			"enabled": true
-		},
-		{
-			"key": "newUsername",
-			"value": "Testing",
-			"enabled": true
-		},
-		{
-			"key": "newPassword",
-			"value": "secret",
-			"enabled": true
-		},
-		{
-			"key": "newFullname",
-			"value": "Testing Account",
-			"enabled": true
-		},
-		{
-			"key": "badUserPayloads",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "currentBadUserPayload",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "refreshToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "badAuthPayloads",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "currentBadAuthPayload",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "newPlaylistName",
-			"value": "Lagu santai",
-			"enabled": true
-		},
-		{
-			"key": "userAToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userAUsername",
-			"value": "userA",
-			"enabled": true
-		},
-		{
-			"key": "userAPlaylistId",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userBToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userBUsername",
-			"value": "userB",
-			"enabled": true
-		},
-		{
-			"key": "userBPlaylistId",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "songIdToPlaylist",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userXUsername",
-			"value": "userX",
-			"enabled": true
-		},
-		{
-			"key": "userXPlaylistId",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userXToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userYUsername",
-			"value": "userY",
-			"enabled": true
-		},
-		{
-			"key": "userXFullname",
-			"value": "User X",
-			"enabled": true
-		},
-		{
-			"key": "userYUserId",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userYToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userYFullname",
-			"value": "User Y",
-			"enabled": true
-		},
-		{
-			"key": "userExportUsername",
-			"value": "User Export",
-			"enabled": true
-		},
-		{
-			"key": "userExportToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "userExportPlaylistId",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "targetEmail",
-			"value": "example@gmail.com",
-			"enabled": true
-		},
-		{
-			"key": "userGenericToken",
-			"value": "",
-			"enabled": true
-		},
-		{
-			"key": "pictureUrl",
-			"value": "",
-			"enabled": true
-		}
-	],
-	"_postman_variable_scope": "environment",
-	"_postman_exported_at": "2021-06-17T06:25:29.474Z",
-	"_postman_exported_using": "Postman/8.6.1-210616-1200"
-}
