+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "open-music-api",
   "version": "1.0.0",
-  "description": "Open Music API v1",
+  "description": "Open Music API",
   "main": "index.js",
   "scripts": {
-    "start-prod": "node ./src/server.js",
-    "start-dev": "nodemon ./src/server.js",
-    "lint": "eslint .",
+    "start:prod": "NODE_ENV=production node ./src/server.js",
+    "start:dev": "nodemon ./src/server.js",
+    "lint": "eslint ./src",
     "migrate": "node-pg-migrate"
   },
   "author": "Rizky Sam Pratama",
@@ -16,23 +16,18 @@
   },
   "license": "ISC",
   "devDependencies": {
-    "eslint": "^7.21.0",
+    "eslint": "^7.24.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
     "nodemon": "^2.0.7"
   },
   "dependencies": {
-    "@hapi/hapi": "^20.1.0",
-    "@hapi/inert": "^6.0.3",
-    "@hapi/jwt": "^2.0.1",
-    "amqplib": "^0.8.0",
-    "aws-sdk": "^2.939.0",
-    "bcrypt": "^5.0.1",
-    "dotenv": "^10.0.0",
-    "joi": "^17.4.0",
-    "nanoid": "^3.1.20",
-    "node-pg-migrate": "^5.9.0",
-    "pg": "^8.6.0",
-    "redis": "^3.1.2"
+    "@hapi/hapi": "^21.3.2",
+    "dotenv": "^16.4.5",
+    "joi": "^17.13.3",
+    "nanoid": "^3.3.6",
+    "node-pg-migrate": "^7.5.2",
+    "nodemon": "^3.0.1",
+    "pg": "^8.12.0"
   }
 }
