--- readme.md
@@ -1,16 +1,0 @@
-# Submission Backend Fundamental Dicoding
-
-## Setup
-
-```
-git clone
-```
-```
-npm install
-```
-```
-npm run migrate up
-```
-```
-npm run start
-```
