--- src/api/authentications/routes.js
@@ -1,19 +1,0 @@
-const routes = (handler) => [
-    {
-        method: 'POST',
-        path: '/authentications',
-        handler: handler.postAuthenticationHandler,
-    },
-    {
-        method: 'PUT',
-        path: '/authentications',
-        handler: handler.putAuthenticationHandler,
-    },
-    {
-        method: 'DELETE',
-        path: '/authentications',
-        handler: handler.deleteAuthenticationHandler,
-    },
-];
-
-module.exports = routes;
