--- src/api/collaborations/routes.js
@@ -1,20 +1,0 @@
-const routes = (handler) => [
-    {
-        method: 'POST',
-        path: '/collaborations',
-        handler: handler.postCollaborationHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-    {
-        method: 'DELETE',
-        path: '/collaborations',
-        handler: handler.deleteCollaborationHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-];
-
-module.exports = routes;
