--- src/api/exports/routes.js
@@ -1,12 +1,0 @@
-const routes = (handler) => [
-    {
-        method: 'POST',
-        path: '/exports/playlists/{playlistId}',
-        handler: handler.postExportPlaylistsHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-];
-
-module.exports = routes;
