--- src/api/playlists/routes.js
@@ -1,52 +1,0 @@
-const routes = (handler) => [
-    {
-        method: 'POST',
-        path: '/playlists',
-        handler: handler.postPlaylistHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-    {
-        method: 'GET',
-        path: '/playlists',
-        handler: handler.getPlaylistHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-    {
-        method: 'DELETE',
-        path: '/playlists/{playlistId}',
-        handler: handler.deletePlaylistHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-    {
-        method: 'POST',
-        path: '/playlists/{playlistId}/songs',
-        handler: handler.postSongToPlaylistHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-    {
-        method: 'GET',
-        path: '/playlists/{playlistId}/songs',
-        handler: handler.getSongFromPlaylistHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    },
-    {
-        method: 'DELETE',
-        path: '/playlists/{playlistId}/songs',
-        handler: handler.deleteSongFromPlaylistHandler,
-        options: {
-            auth: 'songsapp_jwt',
-        },
-    }
-];
-
-module.exports = routes;
