+++ src/api/songs/handler.js
@@ -1,3 +1,5 @@
+const ClientError = require('../../exceptions/ClientError');
+
 class SongsHandler {
     constructor(service, validator) {
         this._service = service;
@@ -39,11 +41,27 @@ class SongsHandler {
             response.code(201);
             return response;
         } catch (error) {
-            return error;
+            if (error instanceof ClientError) {
+                const response = h.response({
+                    status: 'fail',
+                    message: error.message,
+                });
+                response.code(error.statusCode);
+                return response;
+            }
+
+            // Server ERROR!
+            const response = h.response({
+                status: 'error',
+                message: 'Maaf, terjadi kegagalan pada server kami.',
+            });
+            response.code(500);
+            console.error(error);
+            return response;
         }
     }
 
-    async getSongsHandler() {
+    async getSongsHandler(request, h) {
         try {
             const songs = await this._service.getSongs();
             return {
@@ -53,11 +71,18 @@ class SongsHandler {
                 },
             };
         } catch (error) {
-            return error;
+            // Server ERROR!
+            const response = h.response({
+                status: 'error',
+                message: 'Maaf, terjadi kegagalan pada server kami.',
+            });
+            response.code(500);
+            console.error(error);
+            return response;
         }
     }
 
-    async getSongByIdHandler(request) {
+    async getSongByIdHandler(request, h) {
         try {
             const { id } = request.params;
 
@@ -70,11 +95,27 @@ class SongsHandler {
                 },
             };
         } catch (error) {
-            return error;
+            if (error instanceof ClientError) {
+                const response = h.response({
+                    status: 'fail',
+                    message: error.message,
+                });
+                response.code(error.statusCode);
+                return response;
+            }
+
+            // Server ERROR!
+            const response = h.response({
+                status: 'error',
+                message: 'Maaf, terjadi kegagalan pada server kami.',
+            });
+            response.code(500);
+            console.error(error);
+            return response;
         }
     }
 
-    async putSongByIdHandler(request) {
+    async putSongByIdHandler(request, h) {
         try {
             this._validator.validateSongPayload(request.payload);
             const { id } = request.params;
@@ -86,11 +127,27 @@ class SongsHandler {
                 message: 'Lagu berhasil diperbarui',
             };
         } catch (error) {
-            return error;
+            if (error instanceof ClientError) {
+                const response = h.response({
+                    status: 'fail',
+                    message: error.message,
+                });
+                response.code(error.statusCode);
+                return response;
+            }
+
+            // Server ERROR!
+            const response = h.response({
+                status: 'error',
+                message: 'Maaf, terjadi kegagalan pada server kami.',
+            });
+            response.code(500);
+            console.error(error);
+            return response;
         }
     }
 
-    async deleteSongByIdHandler(request) {
+    async deleteSongByIdHandler(request, h) {
         try {
             const { id } = request.params;
             await this._service.deleteSongById(id);
@@ -99,7 +156,23 @@ class SongsHandler {
                 message: 'Lagu berhasil dihapus',
             };
         } catch (error) {
-            return error;
+            if (error instanceof ClientError) {
+                const response = h.response({
+                    status: 'fail',
+                    message: error.message,
+                });
+                response.code(error.statusCode);
+                return response;
+            }
+
+            // Server ERROR!
+            const response = h.response({
+                status: 'error',
+                message: 'Maaf, terjadi kegagalan pada server kami.',
+            });
+            response.code(500);
+            console.error(error);
+            return response;
         }
     }
 }
