--- src/api/uploads/index.js
@@ -1,11 +1,0 @@
-const UploadsHandler = require('./handler');
-const routes = require('./routes');
-
-module.exports = {
-    name: 'uploads',
-    version: '1.0.0',
-    register: async (server, { service, validator }) => {
-        const uploadsHandler = new UploadsHandler(service, validator);
-        server.route(routes(uploadsHandler));
-    },
-};
