--- src/api/uploads/routes.js
@@ -1,17 +1,0 @@
-const routes = (handler) => [
-    {
-        method: 'POST',
-        path: '/upload/pictures',
-        handler: handler.postUploadImageHandler,
-        options: {
-            payload: {
-                allow: 'multipart/form-data',
-                multipart: true,
-                output: 'stream',
-                maxBytes: 1000 * 100 * 5,
-            },
-        },
-    },
-];
-
-module.exports = routes;
