--- src/api/users/index.js
@@ -1,11 +1,0 @@
-const UsersHandler = require('./handler');
-const routes = require('./routes');
-
-module.exports = {
-    name: 'users',
-    version: '1.0.0',
-    register: async (server, { service, validator }) => {
-        const usersHandler = new UsersHandler(service, validator);
-        server.route(routes(usersHandler));
-    },
-};
