--- src/api/users/routes.js
@@ -1,19 +1,0 @@
-const routes = (handler) => [
-    {
-        method: 'POST',
-        path: '/users',
-        handler: handler.postUserHandler,
-    },
-    {
-        method: 'GET',
-        path: '/users/{id}',
-        handler: handler.getUserByIdHandler,
-    },
-    {
-        method: 'GET',
-        path: '/users',
-        handler: handler.getUsersByUsernameHandler,
-    },
-];
-
-module.exports = routes;
