--- src/exceptions/AuthenticationError.js
@@ -1,10 +1,0 @@
-const ClientError = require('./ClientError');
-
-class AuthenticationError extends ClientError {
-    constructor(message) {
-        super(message, 401);
-        this.name = 'AuthenticationError';
-    }
-}
-
-module.exports = AuthenticationError;
