--- src/exceptions/AuthorizationError.js
@@ -1,10 +1,0 @@
-const ClientError = require('./ClientError');
-
-class AuthorizationError extends ClientError {
-    constructor(message) {
-        super(message, 403);
-        this.name = 'AuthorizationError';
-    }
-}
-
-module.exports = AuthorizationError;
