+++ src/server.js
@@ -1,52 +1,11 @@
 require('dotenv').config();
 const Hapi = require('@hapi/hapi');
-const Jwt = require('@hapi/jwt');
-const Inert = require('@hapi/inert');
-
-const ClientError = require('./exceptions/ClientError');
-
 const songs = require('./api/songs');
 const SongsService = require('./services/postgres/SongsService');
 const SongsValidator = require('./validator/songs');
 
-const users = require('./api/users');
-const UsersService = require('./services/postgres/UsersService');
-const UsersValidator = require('./validator/users');
-
-const authentications = require('./api/authentications');
-const AuthenticationsService = require('./services/postgres/AuthenticationsService');
-const AuthenticationsValidator = require('./validator/authentications');
-const TokenManager = require('./tokenize/TokenManager');
-
-const playlists = require('./api/playlists');
-const PlaylistsService = require('./services/postgres/PlaylistsService');
-const PlaylistValidator = require('./validator/playlists');
-
-const collaborations = require('./api/collaborations');
-const CollaborationsService = require('./services/postgres/CollaborationsService');
-const CollaborationsValidator = require('./validator/collaborations');
-
-const _exports = require('./api/exports');
-const ProducerService = require('./services/rabbitmq/ProducerService');
-const ExportsValidator = require('./validator/exports');
-
-const uploads = require('./api/uploads');
-const StorageService = require('./services/S3/StoragesService');
-const UploadsValidator = require('./validator/uploads');
-
-const CacheService = require('./services/redis/CacheService');
-
 const init = async () => {
-    const cacheService = new CacheService();
     const songsService = new SongsService();
-    const usersService = new UsersService();
-    const authenticationsService = new AuthenticationsService();
-    const collaborationsService = new CollaborationsService(cacheService);
-    const playlistsService = new PlaylistsService(
-        collaborationsService,
-        cacheService
-    );
-    const storagesService = new StorageService();
 
     const server = Hapi.server({
         port: process.env.PORT,
@@ -58,115 +17,13 @@ const init = async () => {
         },
     });
 
-    // registrasi plugin eksternal
-    await server.register([
-        {
-            plugin: Jwt,
+    await server.register({
+        plugin: songs,
+        options: {
+            service: songsService,
+            validator: SongsValidator,
         },
-        {
-            plugin: Inert,
-        },
-    ]);
-
-    // mendefinisikan strategy autentikasi jwt
-    server.auth.strategy('songsapp_jwt', 'jwt', {
-        keys: process.env.ACCESS_TOKEN_KEY,
-        verify: {
-            aud: false,
-            iss: false,
-            sub: false,
-            maxAgeSec: process.env.ACCESS_TOKEN_AGE,
-        },
-        validate: (artifacts) => ({
-            isValid: true,
-            credentials: {
-                id: artifacts.decoded.payload.id,
-            },
-        }),
     });
-
-    server.ext('onPreResponse', (request, h) => {
-        // mendapat konteks response dari request
-        const { response } = request;
-
-        if (response instanceof ClientError) {
-            const newResponse = h.response({
-                status: 'fail',
-                message: response.message,
-            });
-            newResponse.code(response.statusCode);
-            return newResponse;
-        }
-
-        // Server ERROR!
-        if (response instanceof Error) {
-            const newResponse = h.response({
-                status: 'error',
-                message: response.output.payload.message,
-            });
-            newResponse.code(response.output.statusCode);
-            return newResponse;
-        }
-
-        // jika bukan ClientError, lanjutkan dengan response sebelumnya (tanpa terintervensi)
-        return response.continue || response;
-    });
-
-    await server.register([
-        {
-            plugin: songs,
-            options: {
-                service: songsService,
-                validator: SongsValidator,
-            },
-        },
-        {
-            plugin: users,
-            options: {
-                service: usersService,
-                validator: UsersValidator,
-            },
-        },
-        {
-            plugin: authentications,
-            options: {
-                authenticationsService,
-                usersService,
-                tokenManager: TokenManager,
-                validator: AuthenticationsValidator,
-            },
-        },
-        {
-            plugin: playlists,
-            options: {
-                service: playlistsService,
-                validator: PlaylistValidator,
-            },
-        },
-        {
-            plugin: collaborations,
-            options: {
-                collaborationsService,
-                playlistsService,
-                validator: CollaborationsValidator,
-            },
-        },
-        {
-            plugin: _exports,
-            options: {
-                service: ProducerService,
-                validator: ExportsValidator,
-                playlistsService,
-            },
-        },
-        {
-            plugin: uploads,
-            options: {
-                service: storagesService,
-                validator: UploadsValidator,
-            },
-        },
-    ]);
 
     await server.start();
     console.log(`Server berjalan pada ${server.info.uri}`);
