+++ src/services/postgres/SongsService.js
@@ -13,9 +13,10 @@ class SongsService {
     async addSong({ title, year, performer, genre, duration }) {
         const id = `song-${nanoid(16)}`;
         const insertedAt = new Date().toISOString();
+        const updatedAt = insertedAt;
 
         const query = {
-            text: 'INSERT INTO songs VALUES($1, $2, $3, $4, $5, $6, $7, $7) RETURNING id',
+            text: 'INSERT INTO songs VALUES($1, $2, $3, $4, $5, $6, $7, $8) RETURNING id',
             values: [
                 id,
                 title,
@@ -24,6 +25,7 @@ class SongsService {
                 genre,
                 duration,
                 insertedAt,
+                updatedAt,
             ],
         };
 
@@ -37,7 +39,7 @@ class SongsService {
     }
 
     async getSongs() {
-        const songs = await this._pool.query('SELECT id, title, performer FROM songs');
+        const songs = await this._pool.query('SELECT * FROM songs');
         return songs.rows.map(mapDBToModel);
     }
 
@@ -48,7 +50,7 @@ class SongsService {
         };
         const song = await this._pool.query(query);
 
-        if (!song.rowCount) {
+        if (!song.rows.length) {
             throw new NotFoundError('Lagu tidak ditemukan');
         }
 
@@ -64,7 +66,7 @@ class SongsService {
 
         const result = await this._pool.query(query);
 
-        if (!result.rowCount) {
+        if (!result.rows.length) {
             throw new NotFoundError(
                 'Gagal memperbarui lagu. Id tidak ditemukan',
             );
@@ -79,7 +81,7 @@ class SongsService {
 
         const result = await this._pool.query(query);
 
-        if (!result.rowCount) {
+        if (!result.rows.length) {
             throw new NotFoundError(
                 'Gagal menghapus lagu. Id tidak ditemukan',
             );
