--- src/services/rabbitmq/ProducerService.js
@@ -1,20 +1,0 @@
-const amqp = require('amqplib');
-
-const ProducerService = {
-    sendMessage: async (queue, payload) => {
-        const connection = await amqp.connect(process.env.RABBITMQ_SERVER);
-        const channel = await connection.createChannel();
-
-        await channel.assertQueue(queue, {
-            durable: true,
-        });
-
-        await channel.sendToQueue(queue, Buffer.from(payload));
-
-        setTimeout(() => {
-            connection.close();
-        }, 1000);
-    },
-};
-
-module.exports = ProducerService;
