--- src/validator/authentications/index.js
@@ -1,29 +1,0 @@
-const InvariantError = require('../../exceptions/InvariantError');
-const {
-    PostAuthenticationPayloadSchema,
-    PutAuthenticationPayloadSchema,
-    DeleteAuthenticationPayloadSchema,
-} = require('./schema');
-
-const AuthenticationsValidator = {
-    validatePostAuthenticationPayload: (payload) => {
-        const validationResult = PostAuthenticationPayloadSchema.validate(payload);
-        if (validationResult.error) {
-            throw new InvariantError(validationResult.error.message);
-        }
-    },
-    validatePutAuthenticationPayload: (payload) => {
-        const validationResult = PutAuthenticationPayloadSchema.validate(payload);
-        if (validationResult.error) {
-            throw new InvariantError(validationResult.error.message);
-        }
-    },
-    validateDeleteAuthenticationPayload: (payload) => {
-        const validationResult = DeleteAuthenticationPayloadSchema.validate(payload);
-        if (validationResult.error) {
-            throw new InvariantError(validationResult.error.message);
-        }
-    },
-};
-
-module.exports = AuthenticationsValidator;
