--- src/validator/collaborations/schema.js
@@ -1,8 +1,0 @@
-const Joi = require('joi');
-
-const CollaborationPayloadaSchema = Joi.object({
-    playlistId: Joi.string().required(),
-    userId: Joi.string().required(),
-});
-
-module.exports = { CollaborationPayloadaSchema };
