--- src/validator/exports/schema.js
@@ -1,7 +1,0 @@
-const Joi = require('joi');
-
-const ExportSongsPayloadSchema = Joi.object({
-    targetEmail: Joi.string().email({ tlds: true }).required(),
-});
-
-module.exports = ExportSongsPayloadSchema;
