--- src/validator/playlists/schema.js
@@ -1,14 +1,0 @@
-const Joi = require('joi');
-
-const PlaylistPayloadSchema = Joi.object({
-    name: Joi.string().required(),
-});
-
-const AddSongPayloadSchema = Joi.object({
-    songId: Joi.string().required(),
-});
-
-module.exports = {
-    PlaylistPayloadSchema,
-    AddSongPayloadSchema,
-};
