+++ src/validator/songs/schema.js
@@ -2,8 +2,7 @@ const Joi = require('joi');
 
 const SongPayloadSchema = Joi.object({
     title: Joi.string().required(),
-    year: Joi.number().integer().min(1900).max(2030)
-        .required(),
+    year: Joi.number().integer().min(1900).max(2030).required(),
     performer: Joi.string().required(),
     genre: Joi.string().required(),
     duration: Joi.number().integer().required(),
