--- src/validator/uploads/schema.js
@@ -1,16 +1,0 @@
-const Joi = require('joi');
-
-const ImageHeadersSchema = Joi.object({
-    'content-type': Joi.string()
-        .valid(
-            'image/apng',
-            'image/avif',
-            'image/gif',
-            'image/jpeg',
-            'image/png',
-            'image/webp'
-        )
-        .required(),
-}).unknown();
-
-module.exports = ImageHeadersSchema;
