--- src/validator/users/index.js
@@ -1,14 +1,0 @@
-const InvariantError = require('../../exceptions/InvariantError');
-const { UserPayloadSchema } = require('./shema');
-
-const UsersValidator = {
-    validateUserPayload: (payload) => {
-        const validationResult = UserPayloadSchema.validate(payload);
-
-        if (validationResult.error) {
-            throw new InvariantError(validationResult.error.message);
-        }
-    },
-};
-
-module.exports = UsersValidator;
