--- src/validator/users/shema.js
@@ -1,9 +1,0 @@
-const Joi = require('joi');
-
-const UserPayloadSchema = Joi.object({
-    username: Joi.string().required(),
-    password: Joi.string().min(3).required(),
-    fullname: Joi.string().required(),
-});
-
-module.exports = { UserPayloadSchema };
