--- text-small.txt
@@ -1,1 +1,0 @@
-Hello :)
